Add project identifier to query
